Migration to CMake

+++ includes/utils.h
@@ -7,6 +7,6 @@ long long* read_file(FILE* file, long long* file_size); // function reading arra
 void write_file(FILE* file, long long* source, long long count); // function writing array of integers to file
 void get_random_file(FILE* file, long long count, unsigned long long max_random); // function writing array of random integers to file
 
-int check_dataset(long long* dataset, long long dataset_size); // function checking dataset for null numbers
+short check_dataset(long long* dataset, long long dataset_size); // function checking dataset for null numbers
 
 clock_t benchmark(long long* arr, long long size, void (*algorithm)(long long*, long long, short), short reverse_bit); // function for benchmarking algorithms
